introduce base webview controller, let help webview inherit from it
